+++ ddl/db_legacy_test.go
@@ -47,10 +47,8 @@ import (
 	"github.com/pingcap/tidb/sessionctx/variable"
 	"github.com/pingcap/tidb/store/mockstore"
 	"github.com/pingcap/tidb/table"
-	"github.com/pingcap/tidb/table/tables"
 	"github.com/pingcap/tidb/tablecodec"
 	ntestkit "github.com/pingcap/tidb/testkit"
-	"github.com/pingcap/tidb/types"
 	"github.com/pingcap/tidb/util"
 	"github.com/pingcap/tidb/util/admin"
 	"github.com/pingcap/tidb/util/dbterror"
@@ -199,22 +197,6 @@ func testGetSchemaByName(c *C, ctx sessionctx.Context, db string) *model.DBInfo 
 func (s *testDBSuite) testGetTable(c *C, name string) table.Table {
 	ctx := s.s.(sessionctx.Context)
 	return testGetTableByName(c, ctx, s.schemaName, name)
-}
-
-func backgroundExec(s kv.Storage, sql string, done chan error) {
-	se, err := session.CreateSession4Test(s)
-	if err != nil {
-		done <- errors.Trace(err)
-		return
-	}
-	defer se.Close()
-	_, err = se.Execute(context.Background(), "use test_db")
-	if err != nil {
-		done <- errors.Trace(err)
-		return
-	}
-	_, err = se.Execute(context.Background(), sql)
-	done <- errors.Trace(err)
 }
 
 func backgroundExecT(s kv.Storage, sql string, done chan error) {
@@ -1060,74 +1042,6 @@ func (s *testDBSuite5) TestCheckConvertToCharacter(c *C) {
 	tk.MustGetErrCode("alter table t modify column a varchar(10) charset utf8mb4 collate utf8mb4_bin", errno.ErrUnsupportedDDLOperation)
 	tk.MustGetErrCode("alter table t modify column a varchar(10) charset latin1 collate latin1_bin", errno.ErrUnsupportedDDLOperation)
 	c.Assert(t.Cols()[0].Charset, Equals, "binary")
-}
-
-func (s *testDBSuite4) TestAddColumn2(c *C) {
-	tk := testkit.NewTestKit(c, s.store)
-	tk.MustExec("use test_db")
-	tk.MustExec("drop table if exists t1")
-	tk.MustExec("create table t1 (a int key, b int);")
-	defer tk.MustExec("drop table if exists t1, t2")
-
-	originHook := s.dom.DDL().GetHook()
-	defer s.dom.DDL().SetHook(originHook)
-	hook := &ddl.TestDDLCallback{}
-	var writeOnlyTable table.Table
-	hook.OnJobRunBeforeExported = func(job *model.Job) {
-		if job.SchemaState == model.StateWriteOnly {
-			writeOnlyTable, _ = s.dom.InfoSchema().TableByID(job.TableID)
-		}
-	}
-	s.dom.DDL().SetHook(hook)
-	done := make(chan error, 1)
-	// test transaction on add column.
-	go backgroundExec(s.store, "alter table t1 add column c int not null", done)
-	err := <-done
-	c.Assert(err, IsNil)
-
-	tk.MustExec("insert into t1 values (1,1,1)")
-	tk.MustQuery("select a,b,c from t1").Check(testkit.Rows("1 1 1"))
-
-	// mock for outdated tidb update record.
-	c.Assert(writeOnlyTable, NotNil)
-	ctx := context.Background()
-	err = tk.Se.NewTxn(ctx)
-	c.Assert(err, IsNil)
-	oldRow, err := tables.RowWithCols(writeOnlyTable, tk.Se, kv.IntHandle(1), writeOnlyTable.WritableCols())
-	c.Assert(err, IsNil)
-	c.Assert(len(oldRow), Equals, 3)
-	err = writeOnlyTable.RemoveRecord(tk.Se, kv.IntHandle(1), oldRow)
-	c.Assert(err, IsNil)
-	_, err = writeOnlyTable.AddRecord(tk.Se, types.MakeDatums(oldRow[0].GetInt64(), 2, oldRow[2].GetInt64()), table.IsUpdate)
-	c.Assert(err, IsNil)
-	tk.Se.StmtCommit()
-	err = tk.Se.CommitTxn(ctx)
-	c.Assert(err, IsNil)
-
-	tk.MustQuery("select a,b,c from t1").Check(testkit.Rows("1 2 1"))
-
-	// Test for _tidb_rowid
-	var re *testkit.Result
-	tk.MustExec("create table t2 (a int);")
-	hook.OnJobRunBeforeExported = func(job *model.Job) {
-		if job.SchemaState != model.StateWriteOnly {
-			return
-		}
-		// allow write _tidb_rowid first
-		tk.MustExec("set @@tidb_opt_write_row_id=1")
-		tk.MustExec("begin")
-		tk.MustExec("insert into t2 (a,_tidb_rowid) values (1,2);")
-		re = tk.MustQuery(" select a,_tidb_rowid from t2;")
-		tk.MustExec("commit")
-
-	}
-	s.dom.DDL().SetHook(hook)
-
-	go backgroundExec(s.store, "alter table t2 add column b int not null default 3", done)
-	err = <-done
-	c.Assert(err, IsNil)
-	re.Check(testkit.Rows("1 2"))
-	tk.MustQuery("select a,b,_tidb_rowid from t2").Check(testkit.Rows("1 3 2"))
 }
 
 func (s *testDBSuite4) TestIfNotExists(c *C) {
